refactor: make some code composition changes
- renaming controllers to services (will make them less reliant on the res,req props)
- convert callback db functions to async
