+++ server/db/mongo/index.js
@@ -51,29 +51,30 @@ const init = async () => {
 
 // this pattern may not work for all db types...but using it for now
 // I also don't particularly like the callback pattern
-const getTransactions = (filter, successCallback, errorCallback) => {
-    Transaction.find(filter, { name: 1, amount: 1, date: 1 }, (err, transactions) => {
+const getTransactions = async (filter) => {
+    try {
+        const transactions = await Transaction.find(filter, { name: 1, amount: 1, date: 1 });
+        return transactions;
+    } catch (err) {
         if (err) {
             logger.error("An error occurred while getting transactions: " + err);
-            errorCallback(err);
-            return;
+            return null;
         }
-        successCallback(transactions);
-    });
+    }
 };
 
-const createTransaction = (transaction, successCallback, errorCallback) => {
+const createTransaction = async (transaction) => {
     const newTransaction = new Transaction({ ...transaction });
-    newTransaction.save().then(savedTransaction => {
-        // if savedTransaction returned is the same as newTransaction then saved successfully
-        if (savedTransaction === newTransaction) {
-            logger.info("Transaction added successfully");
-            successCallback(savedTransaction);
-        } else {
-            logger.error("Failed to add transaction");
-            errorCallback();
-        }
-    });
+    const savedTransaction = await newTransaction.save();
+
+    // if savedTransaction returned is the same as newTransaction then saved successfully
+    if (savedTransaction === newTransaction) {
+        logger.info("Transaction added successfully");
+        return savedTransaction;
+    } else {
+        logger.error("Failed to add transaction");
+        return null;
+    }
 };
 
 const createTransactions = async (transactions) => {
@@ -87,26 +88,16 @@ const createTransactions = async (transactions) => {
     return { success: true, savedTransactions };
 };
 
-const getAllTransactions = (successCallback, errorCallback) => {
-    Transaction.find({}, (err, transactions) => {
+const getAllTransactions = async () => {
+    try {
+        const transactions = await Transaction.find({});
+        return transactions;
+    } catch (err) {
         if (err) {
-            logger.error("An error occurred while getting all transactions: " + err);
-            errorCallback(err);
-            return;
+            logger.error("An error occurred while getting transactions: " + err);
+            return null;
         }
-        successCallback(transactions);
-    });
-};
-
-const getAllTransactionAmounts = (successCallback, errorCallback) => {
-    Transaction.find({}, { "amount": 1 }, (err, transactions) => {
-        if (err) {
-            logger.error("An error occurred while getting all transaction amounts: " + err);
-            errorCallback(err);
-            return;
-        }
-        successCallback(transactions.map(t => t.amount));
-    });
+    }
 };
 
 const getAllTransactionCurrencies = async () => {
@@ -142,6 +133,6 @@ const saveCache = (cache) => {
     saveCacheJson(cache);
 };
 
-module.exports = { init, wipeDb, getTransactions, createTransaction, createTransactions, getAllTransactions, getAllTransactionAmounts,
+module.exports = { init, wipeDb, getTransactions, createTransaction, createTransactions, getAllTransactions,
     getCache, saveCache, getAllTransactionCurrencies, getAllTransactionCategories };
