+++ server/routes/price.js
@@ -1,8 +1,8 @@
-var express = require("express");
-var router = express.Router();
-var controller = require("../controllers/Price.js");
+const express = require("express");
+const router = express.Router();
+const service = require("../services/Price.js");
 
 router
-    .get("/:assetType/:currency?", controller.getCurrencyPrice)
+    .get("/:assetType/:currency?", service.getCurrencyPrice)
 
 module.exports = router;
