+++ server/routes/transaction.js
@@ -1,15 +1,15 @@
-var express = require("express");
-var router = express.Router();
-var controller = require("../controllers/Transactions");
+const express = require("express");
+const router = express.Router();
+const service = require("../services/Transactions");
 
 router
-    .get("/:type?", controller.getTransactions) // TODO: update filtering logic
-    .post("/", controller.addTransaction)
-    .post("/all", controller.addTransactions)
-    .post("/csv", controller.processCSV)
-    .post("/md", controller.processMd)
-    .get("/export/:format", controller.exportTransactions)
-    .get("/all/graph", controller.getGraphData);
+    .get("/:type?", service.getTransactions) // TODO: update filtering logic
+    .post("/", service.addTransaction)
+    .post("/all", service.addTransactions)
+    .post("/csv", service.processCSV)
+    .post("/md", service.processMd)
+    .get("/export/:format", service.exportTransactions)
+    .get("/all/graph", service.getGraphData);
 
 // routes POST /, POST /md and POST /csv could probably be consolidated to POST /:format?
